editor updates for label sizing
